utils: Add a generic comparator for sorted sets
Add the utility class `SortedSetComparator` to compare sorted sets
containing comparable elements.
